+++ front/src/app/idea-memos/IdeaMemos.module.scss
@@ -24,38 +24,3 @@
     );
   }
 }
-
-.noDataContainer {
-  height: calc(100vh - var(--back-button-length) - var(--page-padding));
-  width: 100%;
-  padding-bottom: calc(var(--back-button-length) + var(--page-padding) + 80px);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  @media screen and (min-width: 1024px) {
-    height: calc(100vh - var(--back-button-length) - var(--page-padding-lg));
-    padding-bottom: calc(
-      var(--back-button-length) + var(--page-padding-lg) + 160px
-    );
-  }
-}
-
-.message {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 32px;
-}
-
-.svg {
-  height: 64px;
-  width: auto;
-}
-
-.pagination {
-  margin-top: 40px;
-}
