+++ front/src/features/idea-memos/components/IdeaMemos/IdeaMemos.module.scss
@@ -1,0 +1,3 @@
+.pagination {
+  margin-top: 40px;
+}
