+++ front/src/features/idea-memos/components/index.ts
@@ -1,0 +1,4 @@
+import MemosDisplay from "@/features/idea-memos/components/MemosDisplay/MemosDisplay";
+import NoMemoDisplay from "@/features/idea-memos/components/NoMemoDisplay/NoMemoDisplay";
+
+export { MemosDisplay, NoMemoDisplay };
